+++ artemis-gwt/src/main/java/com/artemis/backends/gwt/emu/com/artemis/InvocationStrategy.java
@@ -10,14 +10,16 @@ public class InvocationStrategy extends SystemInvocationStrategy {
 
 	/** Processes all systems in order. */
 	@Override
-	protected void process(Bag<BaseSystem> systems) {
+	protected void process() {
 		Object[] systemsData = systems.getData();
 		for (int i = 0, s = systems.size(); s > i; i++) {
 			if (disabled.get(i))
 				continue;
 
-			((BaseSystem)systemsData[i]).process();
 			updateEntityStates();
+			((BaseSystem)systemsData[i]).process();
 		}
+
+		updateEntityStates();
 	}
 }