+++ artemis/src/main/java/com/artemis/InvocationStrategy.java
@@ -1,7 +1,5 @@
 package com.artemis;
 
-import com.artemis.utils.Bag;
-
 /**
  * Simple sequential invocation strategy.
  * @see SystemInvocationStrategy
@@ -10,14 +8,16 @@ public class InvocationStrategy extends SystemInvocationStrategy {
 
 	/** Processes all systems in order. */
 	@Override
-	protected void process(Bag<BaseSystem> systems) {
+	protected void process() {
 		BaseSystem[] systemsData = systems.getData();
 		for (int i = 0, s = systems.size(); s > i; i++) {
 			if (disabled.get(i))
 				continue;
 
-			systemsData[i].process();
 			updateEntityStates();
+			systemsData[i].process();
 		}
+
+		updateEntityStates();
 	}
 }